requirements.txt
fastapi
uvicorn
requests
youtube-transcript-api
fpdf
python-dotenv
git+https://github.com/mcptools/fastmcp.git